Handle layers events
